+++ CHANGELOG.md
@@ -38,4 +38,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 
+- Modified ncList to use nc-text-input #31 
 - Component nc-list to customize item #96